requirements.txt
Pillow==9.5.0
Pillow==9.5.0
Requests==2.29.0
streamlit==1.22.0
streamlit_lottie==0.0.3
